display appropriate message for invalid session id
