+++ react/src/components/analysis-results.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Row, Col, message, Alert } from "antd";
-import { AnalysisStatus, SERVER_ADDRESS } from "../utils";
+import { AnalysisStatus, SERVER_ADDRESS, getQueryVariable } from "../utils";
 import { Result } from "./result";
 import { Loader } from "./loader";
 
@@ -13,25 +13,28 @@ export class AnalysisResults extends React.Component {
       analysisStatus: null,
       analysisStartTime: null,
       analysisEndTime: null,
-      analysisStatusMessage: null
+      analysisStatusMessage: null,
+      fetchingResult: false
     };
+    this.downloadResult = this.downloadResult.bind(this);
   }
 
   componentDidMount() {
-    const id = AnalysisResults.getQueryVariable("id");
+    const id = getQueryVariable("id");
     if (!id) {
       return;
     }
+    this.setState({ analysisId: id, fetchingResult: true });
     fetch(SERVER_ADDRESS + "status/" + id)
       .then(response => response.json())
       .then(response => {
         this.setState({
-          analysisId: id,
           analysisStatus: response.status,
           analysisProgress: response.progress,
           analysisStartTime: response.start_time,
           analysisEndTime: response.end_time,
-          analysisStatusMessage: response.message
+          analysisStatusMessage: response.message,
+          fetchingResult: false
         });
       });
   }
@@ -47,21 +50,7 @@ export class AnalysisResults extends React.Component {
   }
 
   downloadResult() {
-    window.open(
-      SERVER_ADDRESS + "result/" + AnalysisResults.getQueryVariable("id"),
-      "_blank"
-    );
-  }
-
-  static getQueryVariable(variable) {
-    const vars = window.location.search.substring(1).split("&");
-    for (var i = 0; i < vars.length; i++) {
-      var pair = vars[i].split("=");
-      if (pair[0] == variable) {
-        return pair[1];
-      }
-    }
-    return null;
+    window.open(SERVER_ADDRESS + "result/" + this.state.analysisId, "_blank");
   }
 
   render() {
@@ -76,13 +65,29 @@ export class AnalysisResults extends React.Component {
 
     return (
       <React.Fragment>
-        <Row type="flex" justify="center" style={{paddingTop: '30px'}}>
-          <Col span={8} style={{ textAlign: "center" }}>          
+        <Row type="flex" justify="center" style={{ paddingTop: "90px" }}>
+          <Col span={8} style={{ textAlign: "center" }}>
             {this.state.analysisStatus ? (
               <Result {...progressProps} />
-            ) : this.state.id? (
-              <Loader />
-            ): <Alert type="warning" message="It seems there is a problem with your request. Please make sure the URL is correct." />}
+            ) : this.state.analysisId ? (
+              this.state.fetchingResult ? (
+                <Loader />
+              ) : (
+                <Alert
+                  type="error"
+                  message={
+                    'Session with ID "' +
+                    this.state.analysisId +
+                    '" could not be found. Please make sure the session ID is correct and try again.'
+                  }
+                />
+              )
+            ) : (
+              <Alert
+                type="error"
+                message="It seems there is a problem with your request. Please make sure the URL is correct."
+              />
+            )}
           </Col>
         </Row>
       </React.Fragment>
