+++ react/src/components/result.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { Progress, Button, Icon, Alert, Collapse } from "antd";
 import moment from "moment";
-import {AnalysisStatus} from "../utils"
+import { AnalysisStatus } from "../utils";
+import logo from "../assets/mozi_globe.png";
 
 export const Result = props => {
   const { progress, status, start, end, message, downloadResult } = props;
@@ -15,7 +16,7 @@ export const Result = props => {
 
   return (
     <React.Fragment>
-      <img src="../assets/mozi_globe.png" style={{ width: "100px" }} />
+      <img src={logo} style={{ width: "100px" }} />
       <h2 style={{ marginBottom: "30px" }}>Mozi service results</h2>
 
       {status === AnalysisStatus.ACTIVE && (
