+++ react/src/utils.js
@@ -6,3 +6,14 @@ export const AnalysisStatus = {
 
 export const SERVER_ADDRESS =
   process.env.SERVICE_ADDR || "http://localhost:5000/";
+
+export const getQueryVariable = variable => {
+  const vars = window.location.search.substring(1).split("&");
+  for (var i = 0; i < vars.length; i++) {
+    var pair = vars[i].split("=");
+    if (pair[0] == variable) {
+      return pair[1];
+    }
+  }
+  return null;
+};
